requirements.txt
firebase-admin==6.1.0
python-dotenv==1.0.0
fastapi==0.95.1
uvicorn==0.21.1
pandas==2.0.1